fix: address review feedback (attempt 2) — Add Tavily Search as Parallel Web-Search Option (Serper / DataForSEO)
